refactor: reorganize utils into subfolders
Move IITC bridge layer into utils/bridge/:
- bridge.js -> bridge/index.js
- events-from-iitc.js, events-to-iitc.js, debug-bridge.js -> bridge/

Move plugin/manager layer into utils/manager/:
- manager-service.js, manager-worker.js -> manager/
- plugin-installer.js, plugin-scripts.js, worker-polyfills.js ->
  manager/

Move webview-service.js into existing utils/webview/ subfolder.

Split platform.js into utils/platform/:
- platform/ui.js - layout helpers, insets, list/input fixes
- platform/navigation.js - Android back handler, goBack
- platform/system.js - share, clipboard (native), links, file URI, app
  name
