+++ app/utils/manager/manager-service.js
