+++ app/utils/manager/plugin-installer.js
