+++ app/utils/manager/plugin-scripts.js
