+++ app/utils/manager/worker-polyfills.js
